+++ hardcover/lib/ui/search_dialog.lua
@@ -22,7 +22,10 @@ local HardcoverSearchDialog = InputContainer:extend {
   select_cb = nil,
   title = nil,
   search_callback = nil,
+  left_icon_callback = nil,
+  left_icon = nil,
   search_value = nil,
+  close_callback = nil,
 
   compatibility_mode = true
 }
@@ -52,7 +55,10 @@ function HardcoverSearchDialog:createListItem(book, active_item)
     info = book.users_read_count .. " reads"
   end
 
-  local active = (book.edition_id and book.edition_id == active_item.edition_id) or (book.book_id == active_item.book_id)
+  local active = active_item and (
+    (book.edition_id and book.edition_id == active_item.edition_id) or
+    (book.book_id == active_item.book_id)
+  )
 
   local result = {
     title = title,
@@ -142,6 +148,9 @@ function HardcoverSearchDialog:init()
   if self.search_callback then
     left_icon = "appbar.search"
     left_icon_callback = function() self:search() end
+  elseif self.left_icon_callback then
+    left_icon = self.left_icon
+    left_icon_callback = self.left_icon_callback
   end
   local menu_class = self.compatibility_mode and Menu or SearchMenu
 
@@ -216,6 +225,10 @@ end
 
 function HardcoverSearchDialog:onClose()
   UIManager:close(self)
+  if self.close_callback then
+    self.close_callback()
+  end
+
   return true
 end